cloud/requirements.txt
flask==3.1.1
werkzeug==3.1.3
flask-socketio==5.4.1
pyyaml==6.0.2
gunicorn==23.0.0
networkx==3.5
numpy==2.3.0
gevent==25.5.1
gevent-websocket==0.10.1
requests==2.32.4
python-socketio==5.13.0
python-engineio==4.12.2
psycopg2-binary==2.9.7
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
SQLAlchemy==2.0.30
redis==4.6.0
asyncio-mqtt==0.16.1
asyncpg==0.29.0
transformers==4.53.0
torch==2.7.1
timeout-decorator==0.5.0
tenacity==8.3.0
openai==1.93.0
